Changes to function ringChoice
I converted the switch statement to if/else if statement.  While the custom ring photo does not show up in accordance with the options selected on the form, I am no longer receiving error messages in console.log.  So, I feel like I'm making a step in the right direction...